Rename ModalContents to ModalContentsBase and create new ModalContents
This lets us have the base version in case we need it, but also creates the common version with the extra dismiss button and the header.
