+++ src/unauthenticated-app.exercise.js
@@ -1,23 +1,10 @@
 /** @jsx jsx */
 import {jsx} from '@emotion/core'
-
 import * as React from 'react'
-import VisuallyHidden from '@reach/visually-hidden'
-import {
-  Input,
-  CircleButton,
-  Button,
-  Spinner,
-  FormGroup,
-  ErrorMessage,
-} from 'components/lib'
+
+import {Input, Button, Spinner, FormGroup, ErrorMessage} from 'components/lib'
 import {Logo} from 'components/logo'
-import {
-  Modal,
-  ModalContents,
-  ModalDismissButton,
-  ModalOpenButton,
-} from 'components/modal'
+import {Modal, ModalContents, ModalOpenButton} from 'components/modal'
 import {useAuth} from 'context/auth-context'
 import {useAsync} from 'utils/hooks'
 
@@ -72,19 +59,6 @@ function LoginForm({onSubmit, submitButton}) {
   )
 }
 
-const CircleDismissButton = () => {
-  return (
-    <div css={{display: 'flex', justifyContent: 'flex-end'}}>
-      <ModalDismissButton>
-        <CircleButton>
-          <VisuallyHidden>Close</VisuallyHidden>
-          <span aria-hidden>×</span>
-        </CircleButton>
-      </ModalDismissButton>
-    </div>
-  )
-}
-
 function UnauthenticatedApp() {
   const {login, register} = useAuth()
   return (
@@ -111,10 +85,7 @@ function UnauthenticatedApp() {
           <ModalOpenButton>
             <Button variant="primary">Login</Button>
           </ModalOpenButton>
-          <ModalContents aria-label="Login form">
-            <CircleDismissButton />
-
-            <h3 css={{textAlign: 'center', fontSize: '2em'}}>Login</h3>
+          <ModalContents aria-label="Login form" title="Login">
             <LoginForm
               onSubmit={login}
               submitButton={<Button variant="primary">Login</Button>}
@@ -126,12 +97,9 @@ function UnauthenticatedApp() {
           <ModalOpenButton>
             <Button variant="secondary">Register</Button>
           </ModalOpenButton>
-          <ModalContents aria-label="Registration form">
-            <CircleDismissButton />
-
-            <h3 css={{textAlign: 'center', fontSize: '2em'}}>Register</h3>
+          <ModalContents aria-label="Registration form" title="Register">
             <LoginForm
-              onSubmit={register}
+              onSubmit={login}
               submitButton={<Button variant="secondary">Register</Button>}
             />
           </ModalContents>
